requirements/base.txt
pyserial
boto3
AWSIoTPythonSDK
enum34
pyyaml